move calculation of rates to fedex_rates
